Fixed uncaught exception
Resolved an uncaught exception occurring at run time by the user code. Issue was occurring on line #93. The lift position variable was being initialized by the right lift motors object using its get current position method. This initialization was invalid as the right lift motor object has no value which causes the exception. To resolve this issue the assignment of the variable was moved to the init method on line 123.